+++ src/main/java/com/everypet/auth/oauth2/data/dto/response/GoogleResponse.java
@@ -1,25 +1,22 @@
-package com.everypet.auth.oauth2.dto;
-
-import com.everypet.auth.oauth2.dto.dtoInterface.OAuth2Response;
+package com.everypet.auth.oauth2.data.dto.response;
 
 import java.util.Map;
 
-public class NaverResponse implements OAuth2Response {
+public class GoogleResponse implements OAuth2Response {
 
     private final Map<String, Object> attribute;
 
-    public NaverResponse(Map<String, Object> attribute) {
-        this.attribute = (Map<String, Object>) attribute.get("response");
+    public GoogleResponse(Map<String, Object> attribute) {
+        this.attribute = attribute;
     }
-
     @Override
     public String getProvider() {
-        return "naver";
+        return "google";
     }
 
     @Override
     public String getProviderId() {
-        return attribute.get("id").toString();
+        return attribute.get("sub").toString();
     }
 
     @Override
