+++ src/main/java/com/everypet/auth/oauth2/data/dto/response/NaverResponse.java
@@ -1,24 +1,23 @@
-package com.everypet.auth.oauth2.dto;
-
-import com.everypet.auth.oauth2.dto.dtoInterface.OAuth2Response;
+package com.everypet.auth.oauth2.data.dto.response;
 
 import java.util.Map;
 
-public class GoogleResponse implements OAuth2Response {
+public class NaverResponse implements OAuth2Response {
 
     private final Map<String, Object> attribute;
 
-    public GoogleResponse(Map<String, Object> attribute) {
-        this.attribute = attribute;
+    public NaverResponse(Map<String, Object> attribute) {
+        this.attribute = (Map<String, Object>) attribute.get("response");
     }
+
     @Override
     public String getProvider() {
-        return "google";
+        return "naver";
     }
 
     @Override
     public String getProviderId() {
-        return attribute.get("sub").toString();
+        return attribute.get("id").toString();
     }
 
     @Override
